+++ src/Application/Plugin/Query/Handler/GetPluginQueryHandler.php
@@ -29,8 +29,8 @@ class GetPluginQueryHandler implements QueryHandler
             'author' => $plugin->getAuthor(),
             'active' => $plugin->isActive(),
             'meta' => $plugin->getMeta(),
-            'createdAt' => $plugin->getCreatedAt()->format('Y-m-d H:i:s'),
-            'updatedAt' => $plugin->getUpdatedAt()->format('Y-m-d H:i:s'),
+            'createdAt' => $plugin->getCreatedAt(),
+            'updatedAt' => $plugin->getUpdatedAt(),
         ];
     }
 }